Reimplement app logic
The entire conversions section is now implemented in JavaScript, i.e.,
DOM elements are generated from internal data and displayed on the page
via a single `append` method. The advantage of this over the previous
"fill-in-the-blanks" approach is two-fold:

1. Better performance. Only one other DOM element (aside from `input`)
   needs to be updated, which means less repaints and optimized code.

2. Extensibility. New dimensions, units, and conversion factors can be
   easily added to conversion.js for more conversion options.

Special thanks to Stan Brown's "How to Convert Units of Measurement"
which helped me grok unit conversions and solidify the app structure.
The article can be found at https://brownmath.com/bsci/convert.htm