Refactor number format implementation
`toLocaleString` takes an `option` object as an argument that contains a
`maximumFractionDigits` property which has a default value of 3.

One benefit over `toFixed` is that we get commas (thousands separators)
out of the box. To be safe, we pass "en-US" for the `locales` parameter
because our app does not support other number formats (e.g. in Germany,
the decimal separator is a comma, while the thousands separator is a
period).

Another benefit is that numbers with less than 3 decimal digits will not
have trailing zeroes.