Make the input a "controlled" element
By adding a focus listener and refactoring the `validateInput` handler,
we're always working from the `baseValue` variable so there's a single
source of truth.

When the user clicks (focuses) on the input element, the input value is
replaced by the raw value (without commas and with all decimal digits)
so the user has access to that information. This also prevents the
validation logic from tripping up (otherwise, "3,000" would resolve to
"0").